host_sync_sqlalchemy/requirements.txt
dotenv
ydb[yc]
ydb-sqlalchemy
